snap: add pppd and resolvconf scripts to the snap
These files cannot be accessed from the snap anymore as they are not
included in the core18 base. Note that the resolvconf scripts will be
used only when installed in a UC16 system.

The resolvconf files had to be modified to use the right folders, that
is the reason for including them instead of staging from the debian
package.