Use $SNAP_DATA/conf to store config data
Use $SNAP_DATA/conf instead of $SNAP_DATA/etc/NetworkManager to store
configuration data. The later is the default, but older versions of
the snap used the former, so we need to do the same to keep the
configured connections when the snap is refreshed. Another option
would be to copy the files in the refresh hook.