Flush IPs after running netplan apply
Flush IPs of devices we are taking control of in the wrapper, after
calling netplan apply and before starting NetworkManager. netplan
leaves IPs configured by networkd around and that confuses
NetworkManager. This is a workaround until LP:#1870561 is fixed.  As
we are using the 'ip' command for this, we need to add network-observe
to the list of interfaces we want.