tests: disable some non-working tests
There are some tests that do not run properly due to:

1. A bug in snapd that messes the mount namespace of NM:
https://bugs.launchpad.net/snapd/+bug/1884806
2. NM not reading configuration files created by netplan because those
lack a .nmconnection suffix. This is because the xenial version of
netplan does not add the suffix, while the more modern in focal does
add it.

2) happens because we run the focal NM on a system with xenial
base. 1) might be happening because of that too. We need to move to
test on UC20 and re-evaluate the situation. Unfortunately, that is not
easy as UC20 still lacks a way of customizing the images in a
reasonable way yet.