spread: fix tests to work with epochs
The spread tests used a little known trick of installing
a version of the network-manager snap from the store before
side-loading the snap under test, in order to ensure all of
the interface connections were connected prior to testing.
This no longer works with the epoch introducted for the uc20
version of network-manager due to the switch to netplan based
system connection files.

Signed-off-by: Tony Espy <espy@canonical.com>