Revert "spread: fix tests to work with epochs"
This reverts commit bd2e479614ec595050aaf4f0f63f6aec12adc095.
The change was needed so tests could run after changing the
epoch. However, this caused races as connections happen after the snap
is installed, and NM starts without all the permissions it
needs. Revert this back as now we have an epoch 1 snap already
published in 20/beta and we can install it without errors.