Refresh 0002-nm-netplan-keyfile.patch
netplan 0.103 compat changes are now integrated with the original patch, as
that version landed in UC20, changes to NM's unit-tests are not needed anymore.

Also, avoid calling into the netplan_generate() libnetplan API, this would just
execute the 'netplan generate' CLI, which would then call the
io.netplan.Netplan.Generate() DBus API in snap environments.
We can avoid this additional redirection and be independent from deprecation
or changes that are incompatible with the snap environment in the
netplan_generate() API, by calling the 'netplan generate' CLI directly.
Ideally, in the future we would avoid this redirection as well and call into
the io.netplan.Netplan.Generate() DBus method directly. But we need to pass a
--root-dir argument during the execution of NM's unit-tests, that is currently
not supported by the DBus method, so let's stick with the CLI for now.