Update netplan integration for v0.104 compatibility
netplan_clear_netdefs() is being called to early, therefore netdef_id is
undefined/garbage and can not be used to calculate a path.

This did work in netplan <= 0.103 as there was a memory leak and netdef_id
was not actually cleared, starting with 0.104 the memory leak is fixed and
we need to adopt the NetworkManager integration accordingly.

This change is compatible with old and new versions of libnetplan.