Do not enable for APs SAE if key management is psk
Enabling SAE when creating an access points make is work only when SAE
is supported by the connecting device, which is still not that
common. Disable then SAE if key management is psk: if wpa3 is desired
for the AP, "sae" needs to be set explicitly. Maybe this is a driver
issue and this patch will need to be reverted after consulting with