Make sure scripts in /etc find /sbin/resolvconf
Create a layout for /sbin/resolvconf, as some scripts from /etc check
for its existence to manipulate or not the resolv.conf database. This
has caused issues with pppd if network-manager is installed on a UC16
system, as files from /etc come from the system and not from the snap
base. Race conditions could lead to bad content in resolv.conf in that
case.